requirements.txt
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
xgboost==3.0.0
lightgbm==4.6.0
imbalanced-learn==0.13.0
pyyaml==6.0.2
jupyter==1.1.1
notebook==7.3.2
joblib==1.4.2
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
